move expand_literals from pdr_prop_solver
expand_literals() is used to break down complex literals to allow for
better generalization. For example, 'a=b' is broken into 'a<=b & b<=a'.

this patch moves it from pdr_prop_solver into pdr_util. This gives more
control on which literals to expand and which not.

For now, this is particularly important for equalities. We want some
equalities to be treated as is and some lowered to arithmetic.