[NEW] priority queue of obligations and flxible_trace
Major change in how obligations (model_node) are handled.

Use a priority_queue to keep track of all active obligations (not just
the leaves of the search tree). This changes how the obligations are
managed (i.e., the top-level loop in spacer::context::check_reachability()).

Allow for obligations at different depth (i.e., blocked at a lower level
and pushed to a higher). model_node::depth() is the number of times the
obligation was blocked and moved to a higher level. This is enabled
using fixedpoint.flexible_trace=true command line option.