[FIX] pre-simplifying expressions before asserting
There seems to be some bug with how the simplifier (or assertions in it)
interact with trivial or non-canonical formulas. In a number of cases
the check that a simplification is trivial fails because the expressions
are not syntactically identical.

A work-around is to run the simplifier in non-proof mode before
asserting.

In general, this might be more effective use of the solver. We only need
the proof to extract the interpolant. Simplifications that are done at
the level of the transition relation or individual assertions are
irrelevant.

Currently enabled by default since I did not find a simple way to pass
options to this part of the code.