emulating BMC with SPACER
On hard sat examples spacer gets stuck in propagate().
This commit adds two options:

   pdr.skip_propagate=[true|false] Skip propagate phase completely.
   pdr.max_level=[UINT] Maximum number of levels to explore.

When pdr.skip_propagate=true, propagate phase of PDR is skipped and it
will never converge to 'unsat. This turns SPACER into a BMC engine.

It is useful to combine it with pdr.max_level set to an expect
counter-example depth.