playing with restarts. Off by default.
It's late and I'm just playing around. Thinking that perhaps instead of
resetting the obligation queue between each propagate, we can reset it
when it gets too large. In some sense, the obligation queue is like a
decision stack of a sat-solver. Resetting it is like restarts.

This seems to work nicely, but I need to port other strategies, decide
on good constants, and evaluate of a size-able benchmark set.