simple example to illustrate an issue with arrays
/z3 -v:1 ../examples/dl_smt2/array_eq.smt2 fixedpoint.engine=spacer \
    fixedpoint.xform.inline_eager=false \
    fixedpoint.xform.inline_linear=false \
    fixedpoint.xform.slice=false \
    fixedpoint.use_array_eq_generalizer=false -tr:spacer

Disable inlining and slicing. Otherwise, the example is inlined.

It is easy to turn this into a version that cannot be inlined. Either by
making inlining very costly, or by adding recursion.

With the option

  fixedpoint.use_array_eq_generalizer=false

the solver gets into an infinite loop, generating infinitely many
concrete array facts.

With the option

  fixedpoint.use_array_eq_generalizer=false

the solver gets the array equality.

Still need to cleanup the story of arrays and proof how powerful the
result is. That is, show that for a restricted (but useful) set of Horn
clauses the algorithm is relatively complete. In other words, that it is
an interpolating decision procedure for arrays.