[FIX] flatten input to qe_project
Heuristic projections (such as qe_lite) are based on the assumption that
top-level formula is flat (i.e., one n-ary conjunction). Everything is
then done relative to the top-level conjuncts.

This is not true of the clients of qe_project(). For now, adding an
extra flattening step in the beginning of qe_project(). This requires a
more careful examination to identify all hidden assumptions.

Note that the pretty printer flattens the formula when printing. So this
bug is not apparent from looking at the trace log.