[FIX] update of m_expanded_lvl
Only update m_expanded_lvl whenever new lemmas have been added at that
level. Currently, it was being updated whenever a proof obligation was
asked at that level. However, the level remains clean (not expanded for
propagate purposes) unless a lemma has been generated for the
obligation. If the obligation is Killed or True, the level should not
count as expanded.