option to enable using default smt solver
In the pdr engine, smt-solver was restricted with no bound and equality
propagation. This is potentially necessary for farkas learner. However,
it seems to have a significant performance impact right now.

I have disabled it, and a couple of examples still work (but now much
faster). It might need to be re-enabled. In which case, this is a good
validation case for using multiple differently configured solvers.