itp_solver: external api for proxies
Algorithms, such as mux and maxsat, need to make many calls to the
solver with different sets of assumption literals.

Exposed api to apply assumption literals to external vectors of
expressions, and api to covert assumptions back into their definitions.

This is potentially fragile since the proxy variables might be reused
while they have been exposed to the user.