Working around a mess with smt_params
There is no convenient way to configure a solver object.

Underlying smt::kernel and smt::context are configured by updating their
smt_params objects. However, there is no generic interface to access
smt_params (because it is implementation specific).

The only alternative is to use strings in params_ref and pass them
through solver::updt_params(). But this seems silly for something as
tightly integrated as an engine for a logic.

Furthermore, datalog engine assumes that there is only one copy of
smt_params shared between all solvers, which is limiting.

The current solution is an uneasy compromise. smt_params are exposed
through various hacks. This needs to be re-evaluated once the use cases
for multiple parameters will become clear.