[Spacer] children creation with uncertainty
With weak solver, it is possible to get a model that is not a model for
the lemmas of the child to be created. In this case, we bail out and
bump weakness.

It is also possible to attempt to fix the model instead of giving up on
it completely.