[Spacer] constructing unit resolution
ast_manager provides two API to construct unit resolution proof:
```
  proof* mk_unit_resolution(unsigned num_proofs, proof * const * proofs, expr * new_fact)
  proof* mk_unit_resolution(unsigned num_proofs, proof * const * proofs)
```

The first one requires the client to provide the resolvent, the second
computes the resolved based on the arguments. Unfortunately, the first
API is restricted to the case where new_fact is a unit literal or
false. There is nothing specific in the API, but debug assertions check
for that case.

The second form does a fairly inefficient matching computation.

We fall-back on the second form for now, until the efficiency of that
rule becomes a bottleneck.