(spacer) better handling of array equality in mev
Equality handling in model_evaluator for arrays is really slow.  Nikolaj
had some benchmarks that were failing. I have fixed that a bit. Nikolaj
fixed it in yet another way. I never had a chance to go back and try a
few other alternatives. This code is likely to clash with current Z3
implementation. But, it is what was tested for spacer.