(quic3) fixed memory leak
Skolem variables are created but are not necessarily kept.
It used to work because the constraints are given to the solver that
is never reset. Thus, there is always at least one reference.

In a new mode (not yet committed), the solver is reset and Skolem
variables disappear.

I've added m_skolems field to spacer::context to keep track of all
created skolems.