(quic3) new option fixedpoint.spacer.keep_proxy=[true|false]
The default (and previous default) is 'true'.

When true, the internal solvers are never reset. All queries are done
by enabling appropriate assumptions.

When false, the solver is pushed before each query and pop'ed after.
This ensures that no constraints from one query are available for the
next.

We suspect that additional constraints in the solver might create
extra quantifier instantiations.