treat inductive and relative inductive qlemmas uniformly
A quantified lemma is only added in the pre-state, but not the
post-state of every frame.

A ground lemma is added to both pre- and post-state.
While adding a ground lemma to the post-state is not strictly
necessary (it is entailed from the pre-state), it is considered to be
a useful fact for the solver.

Note that previous behaviour was to add inductive (i.e., F-infinity)
quantified lemmas to both pre- and post-state. Hence, quantified
lemmas were not treated uniformly.

Since there was no particular reason for the difference, I've changed
the code to be more uniform for now.