[quic] new method pred_transformer::is_qblocked()
returns true if a cti can be blocked in the current frame using
quantified lemmas.

Not wired in, this is intended to test how many instantiation
opportunities we have.

It is intended that the solver in is_qblocked() is configured with
mbqi even if solvers used elsewhere are not. I have not checked that
this is in fact the case, but this should not be hard to ensure.

It is expected that mbqi will not diverge, eventhough it diverges
elsewhere.